pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kalimcp"
version = "0.1.0"
description = "Kali Linux MCP Server — expose 500+ security tools to AI via Model Context Protocol"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "KaliMcp Team" }]

dependencies = [
    "mcp[cli]>=1.0.0",
    "pydantic>=2.0",
    "aiofiles>=24.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30",
    "pyjwt>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
kalimcp = "kalimcp.server:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/kalimcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120
